Refactor into mode classes.
To allow more generic use of the different modes (currently, images
and time), there is now an LedServiceMode class.  It can respond
to commands and optionally loop.  A mode is enabled by sending
a message over redis starting with its name; if it is already
the current mode, the rest of the command is sent to the existing
mode object.  Other commands, e.g. `next_mode`, are handled by the
main mode-handler loop.